+++ src/main/java/br/unirio/pm/distancia/IDistanceCalculator.java
@@ -2,7 +2,6 @@ package br.unirio.pm.distancia;
 
 public interface IDistanceCalculator {
 
-	public int distance(String palavra1, String palavra2);
+	public double distance(String palavra1, String palavra2);
 
-	/* public double distance(String palavra1, String palavra2); */
 }
